Fixing issue with multiple triggers at the same time
This commit resolves the issue, that an action is not running, due to 2 actions are triggered at the same time. In "single" mode, the second trigger (in my case often the "packinputpower") is blocked by home assistant. This results, that the value was not changed. In "parallel" mode, both actions will be executed.